+++ Editors/BackgroundEditor.h
@@ -6,25 +6,39 @@
 #include "BackgroundRenderer.h"
 
 #include <QItemDelegate>
-#include <QDebug>
-#include <QObject>
-#include <QWidget>
+#include <QDataWidgetMapper>
+#include <QMetaObject>
+#include <QMetaProperty>
 
-class DataPusher : public QObject
+#include <QDebug>
+
+class ImmediateDataWidgetMapper : public QDataWidgetMapper
 {
 	Q_OBJECT
-
-public:
-	DataPusher(QWidget *editor, QItemDelegate *delegate): QObject(editor),
-		editor(editor), delegate(delegate) {}
-public slots:
+private slots:
 	void widgetChanged() {
+		auto delegate = this->itemDelegate();
+		auto editor = static_cast<QWidget*>(this->sender());
 		delegate->commitData(editor);
 		delegate->closeEditor(editor, QAbstractItemDelegate::SubmitModelCache);
 	}
-private:
-	QWidget* editor;
-	QItemDelegate* delegate;
+
+public:
+	ImmediateDataWidgetMapper(QObject *parent = 0): QDataWidgetMapper(parent) {}
+
+	void addMapping(QWidget *widget, int section, QByteArray propertyName = "") {
+		QDataWidgetMapper::addMapping(widget, section, propertyName);
+		propertyName = this->mappedPropertyName(widget);
+		auto widgetMetaObject = widget->metaObject();
+		auto property = widgetMetaObject->property(widgetMetaObject->indexOfProperty(propertyName));
+		auto metaObject = this->metaObject();
+		QMetaObject::connect(
+			widget,
+			property.notifySignalIndex(),
+			this,
+			metaObject->indexOfSlot("widgetChanged()")
+		);
+	}
 };
 
 namespace Ui {
